Cleaning up project for other to use and start creating pages using react-router-dom
