--- Frontend/anime-search-app/src/screens/Home/main.css
@@ -1,36 +1,0 @@
-.Background {
-    position: relative;
-    background-image: url("https://i.etsystatic.com/27448597/r/il/d9d5b5/2934931410/il_570xN.2934931410_bz9g.jpg");
-    background-size: 300px;
-    width: 100vw;
-    height: 100vh;
-    background-repeat: repeat;
-}
-
-.BackgroundOverlay {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0, 0, 0, 0.75);
-    background-repeat: repeat;
-}
-
-.CenterSearch {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    min-height: 80vh;
-    flex-direction: column;
-}
-
-.LogoText {
-    font-family: "Monoton", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-    color: white;
-    user-select: none;
-}
